Workaround for runtime error with CbmRoot
The changes are a workaround for CbmRoot. With the latest FairSoft versions
on macosx it comes otherwise to runtime errors when JIT compiling a macro.
The problem is a define statement in termios.h which defines B0 to be 0.
This coonflicts with the parameter B0 used in the header file.